holdingsBE/requirements.txt
flask
flask-cors
gunicorn
openai
python-dotenv
faiss-cpu
PyMuPDF
numpy